pyproject.toml
[project]
name = "doc2latex"
version = "0.1.0"
description = "Document to LaTeX converter - 将Word文档转换为LaTeX格式的工具"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Doc2LaTeX Team"}
]
keywords = ["latex", "word", "document", "converter", "中文"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: LaTeX",
    "Topic :: Office/Business :: Office Suites",
]
dependencies = [
    "pylatex>=1.4.2",
    "python-docx>=1.1.2", 
    "treelib>=1.7.0",
    "styleframe>=4.2",
    "opencc-python-reimplemented>=0.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
]
win32 = [
    "pywin32>=306; sys_platform == 'win32'",
]

[project.scripts]
doc2latex = "doc2latex.main:main"

[project.urls]
Homepage = "https://github.com/your-username/doc2latex"
Repository = "https://github.com/your-username/doc2latex"
Documentation = "https://github.com/your-username/doc2latex#readme"
"Bug Tracker" = "https://github.com/your-username/doc2latex/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.sources]
"src" = ""

[tool.black]
line-length = 88
target-version = ['py39']
include = 'src/.*\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | archive
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"
packages = ["doc2latex"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/doc2latex",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
